RFC: Add can.Observe.List.sort() method
I have some jmvc code which I'm porting to canjs.

It needs to sort the model list which in jmvc was just a plain array. Not so in canjs, hence the patch.

I'm not sure what the event name and args should be... so I use
"length" for now.

Please let me know and I will update the patch.