canMakeObserve: strictly allowed types
Made objects for canMakeObserve more strict; only allow plain JS objects
or existing instances of can.Observe or derived classes (such as
can.Observe.List, can.Model, etc. )

Allowing other things to be wrapped into can.Observe instances may cause
problems as well, so it is better to white-list and be conservative
about what is and is not allowed.

(For example; even wrapping can.Control, which is part of CanJS itself
throws.)