Changed the "remove item from nested array" test so that it passes in IE7/8
For whatever reason, IE7/8 will effectively remove the item using
splice (length changes, items shift down) but the value itself doesn't
seem to be guaranteed to be removed. This might be because of accessing
array.1 as a string key instead of as an integer flagging something in
IE's array implementation. This doesn't occur in IE9/10.