Failing test for custom element namespace in IE8
This adds a failing test to the issue where custom elements that have
been registered with can.view.tag are being rendered in IE8 with a
preceding `:` because they lose their namespace. This problem does not
occur in Mustache.

The issue is likely caused by stache using cloneNode internally. There
are various issues with cloneNode that are solved by jQuery and HTML5
Shiv includes a note in their readme that they do not address the
cloneNode problem.