Fix for custom tags in IE8 using stache
This fixes a bug where custom tags used in IE8 would result in the tags
not being part of the document's namespace and therefore would be
rendered like: <:my-tag></:my-tag>.  This would also break for
non-custom tags like <header> and <footer> because stache uses
`cloneNode` which is not fixed by html5shiv.

The fix is a custom `cloneNode` for browsers with this issue that
recursively creates elements as a replacement for the browser's
cloneNode. Fixes #985