Fixes compute.read with a Map wrapped in a compute
Previously if compute.read received a can.Map wrapped in a can.compute
it would attempt to call `.attr` against the compute itself. This is
because we were checking for `prev` to be a compute and if so getting
the value, but not setting the value on `cur` as well.