All jquery pluginify test are now passing
This updates everything needed so that jquery's pluginified test all
pass. The biggest obstacle was that there isn't a central place to get
all of the tests, so I created test/all-tests.js that contain all of the
tests. The old pluginify builder had a way of combining modules that
need to be pluginified into one; the new pluginify doesn't have that.
You must specify a `main` module, which is why all-tests.js was created.