Added jquery to peer dependencies
Changed the verion of the required jquery to a range of 1.9.0 to less than 3.0.0

Kept jquery in dev dependencies so that it can be install at the root level of canjs via npm install

Removed jquery from devDependencies as it is not needed.

Because of https://github.com/npm/npm/issues/6323, peerDependencies was not getting installed at the root level.  This was fixed in v2.1.5 of npm, therefore an engine constraint was added to package.json

Added jquery back to devDependencies

Added it back because only being in peerDependencies is failing the CI because of the npm version of 1.4.28.

Removed engines property

Because we are adding jquery back to devDeps we no longer need to constrain the version of npm