Fix can.scope and autorender
This fixes can.scope and autorender. There were two problems:

1. autorender depended on can.scope to create a scope for an element
when one didn't already exist and this no longer was happening. The fix
was to make can.scope create a scope if there isn't already one. Added a
test for can.scope for this.

2. autorender was setting up scope for the scoped script tag and adding
attributes to the scope but this wasn't working. The reason is that
`el.attributes` is a NamedNodeMap of Attrs which are objects, not
strings. The fix was to use the `attr.name` property.

Fixes #1503