Use localLoader for doing defines
When doing a `System.define` we need to use the localLoader if it's
available. This is because during the build the BuildSystem will be the
one executing the plugin, but we need for those defines to be on the
LocalSystem loader because it is the one that does the dependency
loading.