Pass in document to can.appendChild
This makes it so that everywhere we are appending we pass in the
can.document (if one exists, which is only when using vdom) so that the
"inserted" event will be triggered on elements.

We do this by, in node, waiting for all importPromises to complete and
then appending, which causes all of the inserts to happen.