Binding to element's properties
This adds a can.view.attrValue function that can be used to define
callbacks based on an attribute's value.  This is used to do bindingings
for things like `<div hidden="{hidden}"></div>` and then setting the
property.

This is not ready yet as it doesn't update the value (like a can.Map
		with a `hidden` property) when the element's properties change.

This is related to #1700