Set the promise value when using can-tag
This updates can/view/import to set the `value` property on the
can-import viewModel after the promise has resolved when used with
can-tag. The reason this is needed is because when handing off control
to a can.Component there will be a new viewModel created for that
element. So we need to set the import value so that it can be used by
the can-tag (or within the template as a reference value).