Implement ~ helper
The ~ helper is used to create urls (for `<img>` for example) that are
bound to the root of your application or relative to the current module.

If the lookup is relative, such as:

```
{{~ "../hello/world.svg"}}
```

This will lookup the location relative to the current template module
(if using Steal).

Otherwise it will lookup the location through, in descending importance:

* can.baseUrl
* System.baseUrl (if using Steal/System)
* location.pathname

Fixes #1816

Includes docs