Switch helper: get expression data in switch
This fixes #1857. The bug was that if the expression passed into switch
changed the switch helepr would not re-evaluate and instead on the
switches would re-evaluate. So you could never go from one case to the
default. This fixes it by setting up the binding to the expression
within the switch helper function, so that if the expression changes it
will reevaluate the entire switch.