A lot of the prelim work for #1699 and #1700.
This changes what can.compute.read takes to include
info about how the property should be read.

This changes scope to deal with it.

This adds a basic Helper expression that seems to work like the
old helper expression.

This adds a better expression parser.

I need to hookup the new expressions in place of the old expressions and make sure expression.Call works as well as the new operators.