Docs: Add reference to new bindings in Stache docs
I've watched 3 people look for the docs on view-bindings inside the Stache docs.  This seems like the most-logical place to create a link between the sections.  I put the code example in place so people can recognize what it is, since `can.view.bindings` won't ring a bell for people, even though it might be the best name for them.