Use preventDataBindings in can/component
This uses the element's data `preventDataBindings` to control whether we
set up bindings in can/component. This is needed because in the case of
can-import's `can-tag` functionality we do not want bindings to be set
up. In can/view/import now we toggle preventDataBindings to true so that
can/component will not set them up and then afterwards toggle it back to
false so that can/view/bindings will. Closes #2045