Correctly handle hookup for self-closing tags in can.view.txt
In some cases hookup is done by can.view.txt on tags that are
self-closing. The default hookup logic always renders tags with content,
which is bad because browsers attempt to fix this when parsed into a DOM
fragment and end up bubbling the content outside of the node, leaving
ugly @@!!@@ marker text everywhere.
This fixes that by adding a few checks for self-closing tags and
rendering appropriate hookup placeholder markup.