Wait to add %root to a component’s viewModel data until all of the other attributes have been bound
When a component is set up, `%root` is added to the initial `viewModel` data so it’s available in every scope. However, this causes issues when the `%root` bindings are set up before all of the other properties, so this moves the setup for `%root` until after all the other properties have been bound.

Fixes #2250