Moving all sauce labs tests to test/test-sauce-labs.js
This removes the need for grunt. It also increases the consistency and simplicity of the test-sauce-labs.js script by moving the the localtunnel into the .travis.yml using the sauce_connect addon and moving the testee server code out of the script as well. It also updates can-* dependencies to the latest.