Fix some issues in the TodoMVC guide
The biggest change in this commit is not using the generic API names in the descriptions that appear before the code changes.

Here’s an example of a sentence that appeared previously:

> Use `{{#each expression}}` to loop through every todo.

With this commit, the following is displayed:

> Use `{{#each todos}}` to loop through every todo.

The code still links to the API docs and it makes it easier for readers to connect the sample code with the description.

Related to #2867