Update can-component’s accordion demo to work in 4.0
The demo made use of `leakScope,` which produced buggy behavior in 4.0, so the demo no longer uses it.

Additionally, this demo is used in can-component’s ViewModel docs to show how you can change a component’s element attributes to update the view model, but I don’t think this is supported in 4.0, so the demo now shows using `can-view-model` to get the component’s view model and set a property.