Exit with an error when the deps-bundle script fails
Previously, if you ran the deps-bundle script and steal-tools reported an error, the script would still exit successfully and the next script would run (e.g. when building the site/docs). This would leave you with an out-of-date (or non-existent) dev-bundle.

This fixes the issue by catching any errors reported by steal-tools and exiting the script in a way that will make Node report the error and stop.