Promote steal-stache to being a dependency of can
This will be a minor release. This adds `steal-stache` as a dependency
of can. This prevents the need to separately install steal-stache and
prevents issues when upgrading can but not steal-stache.

After this change users should remove `steal-stache` from their
dependencies and change their package.json from:

```js
{
  ...
  "steal": {
    "plugins": [
      "steal-stache"
    ]
  }
}
```

to:

```js
{
  ...
  "steal": {
    "plugins": [
      "can"
    ]
  }
}
```