Fix SauceLabs badge to only update from canjs/canjs
Previously, any builds that were PRs or not on master would run with the `canjs-not-master` Sauce Labs user so only builds on `master` would be run with the main `canjs` user.

This fixed the Sauce Labs badge for builds within the `canjs` repo, but builds in all the other repos continued to use the `canjs` Sauce Labs user, so they still affected the badge.

This fixes the issue by using the `canjs-not-master` Sauce Labs user only for builds on `master` and using the badge for that user everywhere. Unfortunately, Sauce Labs user names can’t be changed and we can’t add any more users to our main account, so now the `canjs-not-master` user is only used for builds on `master`.

Fixes https://github.com/canjs/canjs/issues/4764